step 5: upgrade `@angular-builders/custom-webpack`
The `@angular-builders/custom-webpack` package does not
support the `ng-update` schematic, thus I had to manually upgrade it.

At this time, when running the application, it should spin up successfully.
However, in the browser, you will only see the first line of text being red.
On inspecting the dom you will recognize, that the css styles have been
loaded successfully, but the font-face src urls have not been adapted.

When building the app, you will recognize, that the font files have not
been copied to the `dist` folder.